Calculated zoom is now adjusted by a constant zoomAdjustment
In order to avoid pins being placed on the edge of the displayed map,
the calculated best zoom is now adjusted by a constant named
zoomAdjustment with a value equal to 1.  Previously a value of .25
was being used which was automatically rounded down to 0.  Therefore,
one of the test points was changed in order to reflect the larger zoom
after adjustment.