Searching now zooms to display target number of pins
When the user searches the map's zoom is now adjusted to include the target
number of pins (currently 1 for groups and 10 for connect) when possible.
However, it is set to have a maximum and minimum zoom of 3 and 15 respectively.
In order to solve the problem with pins appearing right on the edge of the map,
after the proper zoom is calculated, .25 is subtracted in order to zoom out
slightly farther (so the actual range of allowable zooms is 2.75 to 14.75).
We can adjust this extra zoom out later if we desire.

The zoom calculation is handled in google-map.service.ts, and I added tests for
the zoom calculation in google-map.service.spec.ts.  These tests only cover the
other functionality of the GoogleMapService (such as calculating bounding boxes)
indirectly.  It may be a good idea to add tests for those features in the
future.  In order to facilitate testing, I added a public method,
getPinsInBounds, which may be useful in other situations.

The current algorithm for adjusting the zoom starts with a zoom of 9 and then
uses recursion to repeatedly increase or decrease the zoom by 1 until it finds
the maximum zoom that contains the target number of pins.  The starting zoom was
chosen to be halfway between 3 and 15; we may want to look at whether we should
use a different starting zoom.  A more efficient algorithm would be to instead
adjust by half of difference between the current zoom and the min or max zoom
(depending on whether the zoom needs to be adjusted up or down).  In effect, we
would be doing a binary search for the proper zoom.

Finally, I significantly refactored the code in google-map.service.ts.